Fix 64-bit experience display for OTCv8

+++ src/player.cpp
@@ -2108,11 +2108,12 @@ uint16_t Player::getBasisPointLevel(uint64_t count, uint64_t nextLevelCount)
 		return 0;
 	}
 
-	uint16_t result = ((count * 10000.) / nextLevelCount);
-	if (result > 10000) {
+	// Use double precision to avoid overflow with large XP values
+	double result = (static_cast<double>(count) * 10000.0) / static_cast<double>(nextLevelCount);
+	if (result > 10000.0) {
 		return 10000;
 	}
-	return result;
+	return static_cast<uint16_t>(result);
 }
 
 void Player::onBlockHit()
